requirements.txt
torch>=1.8.0
torchvision
numpy
opencv-python
tqdm
Pillow
